Halved the size of pawn score entries. There wasn't a good reason to keep it at 32 bit integers, so I changed it to 16 bits, doesn't seem to run into issues with the reduced width.

+++ Anduril.h
@@ -228,7 +228,7 @@ private:
 
     // finds the pawn structure bonus for the position
     template<bool color>
-    std::pair<int, int> getPawnScore(libchess::Position &board);
+    std::pair<int16_t, int16_t> getPawnScore(libchess::Position &board);
 
     // finds the king safety bonus for the position
     int getKingSafety(libchess::Position &board, libchess::Square whiteKing, libchess::Square blackKing);
